Trigger build workflow on 1st of each month
Artifacts are kept for up to 90 days and then removed.
In datalad/git-annex we download from here, and now it fails since
we are past 90 days since the last build.

Building regularly on cron (once a month) should ensure that we
have an artifact, but ideally we should move distribution of it
elsewhere